Rename yosys binary to avoid path conflicts on windows
