+++ src/main.rs
@@ -1,8 +1,8 @@
 use std::path::Path;
 use clap::Parser;
 
-#[path = "./bin/yosys.rs"]
-mod yosys;
+#[path = "./bin/make_blif.rs"]
+mod make_blif;
 
 #[path = "./bin/read_blif.rs"]
 mod read_blif;
@@ -17,8 +17,8 @@ struct Args {
 fn main() {
     let args = Args::parse();
     let verilog_path = Path::new(&args.src);
-    let blif_buf = yosys::make_blif_path(verilog_path);
+    let blif_buf = make_blif::make_blif_path(verilog_path);
     let blif_path = blif_buf.as_path();
-    yosys::generate_blif(verilog_path, Path::new("res/mc.lib"), blif_path, false);
+    make_blif::generate_blif(verilog_path, Path::new("res/mc.lib"), blif_path, false);
     read_blif::read_blif(blif_path);
 }
